backend/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
python-multipart==0.0.27
pytesseract==0.3.13
pdf2image==1.17.0
pillow==12.2.0
pydantic==2.13.3
python-dotenv==1.2.2
groq==1.2.0
pdf2image==1.17.0